Unified Decrypt tab for both password and auth encrypt
